Changed the elapsedSeconds on GameObjects and GameRound to be a float instead of a double. Most of the math here uses floats anyway.

+++ MarioBros/GameObject.cs
@@ -16,6 +16,6 @@ namespace MarioBros
             VisualDimensions = visualDimensions;
         }
         
-        public abstract void Update(double elapsedSeconds);
+        public abstract void Update(float elapsedSeconds);
     }
 }
